APL-1681 fixed HeightConfig;removed useless constants;

+++ apl-utils/src/main/java/com/apollocurrency/aplwallet/apl/util/env/config/Chain.java
@@ -207,7 +207,7 @@ public class Chain {
         this.initialSupply = initialSupply;
     }
 
-    public long getDecimals() {
+    public int getDecimals() {
         return decimals;
     }
